+++ Options/OptRtPrefix.cpp
@@ -73,14 +73,12 @@ uint8_t TOptRtPrefix::getMetric()
     return Metric;
 }
 
-SPtr<TIPv6Addr> TOptRtPrefix::getPrefix()
-{
+SPtr<TIPv6Addr> TOptRtPrefix::getPrefix() {
     return Prefix;
 }
 
-int TOptRtPrefix::getSize()
-{
-    return 4+22+getSubOptSize();
+size_t TOptRtPrefix::getSize() {
+    return 4 + 22 + getSubOptSize();
 }
 
 std::string TOptRtPrefix::getPlain() {
